Revert "revert"
This reverts commit ee83bd18ab68c2c3c16c8fd51736b780a22544e6.

+++ iguana/util.hpp
@@ -14,8 +14,8 @@
 #include "define.h"
 #include "detail/charconv.h"
 #include "detail/utf.hpp"
+#include "enum_reflection.hpp"
 #include "error_code.h"
-#include "field_reflection.hpp"
 #include "reflection.hpp"
 
 namespace iguana {
@@ -297,8 +297,8 @@ IGUANA_INLINE void write_string_with_escape(const Ch* it, SizeType length,
 
 template <typename T, size_t N>
 IGUANA_INLINE constexpr bool has_duplicate(const std::array<T, N>& arr) {
-  for (size_t i = 0; i < arr.size(); i++) {
-    for (size_t j = i + 1; j < arr.size(); j++) {
+  for (int i = 0; i < arr.size(); i++) {
+    for (int j = i + 1; j < arr.size(); j++) {
       if (arr[i] == arr[j]) {
         return true;
       }
